backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi==0.115.6",
    "google-generativeai==0.8.3",
    "pydantic==2.10.5",
    "python-dotenv==1.0.1",
    "qdrant-client==1.11.3",
    "uvicorn[standard]==0.34.0",
]
